pyproject.toml
[project]
name = "inference-server"
version = "0.1.0"
description = "Kubernetes-native LLM inference platform with multi-tier KV caching"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "prometheus-client>=0.21.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
gateway = [
    "httpx>=0.27.0",
    "sse-starlette>=2.1.0",
]
engine = [
    "httpx>=0.27.0",
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "huggingface-hub>=0.26.0",
    "vllm>=0.6.0",
    "transformers<5.0.0",
]
sidecar = [
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "redis>=5.2.0",
    "huggingface-hub>=0.26.0",
]
gpu = [
    "cupy-cuda12x>=13.0.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]
tracing = [
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-instrumentation-httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared", "data_plane", "control_plane"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring docker-compose services",
]

[tool.ruff]
target-version = "py312"
line-length = 120
extend-exclude = ["Makefile"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
